dao aprte 2

+++ PatronDAODTONM/src/modelo/AbstractDAO/PersonaDAO.java
@@ -2,6 +2,6 @@ package modelo.AbstractDAO;
 
 import modelo.Entity.Persona;
 
-public interface PersonaDAO extends GenericDAO<Persona, Integer> {
+public interface PersonaDAO extends GenericDAO<Persona, Long> {
 	public Persona findByName(String name);
 }
